+++ movie-book/src/main/java/hello/moviebook/book/controller/BookController.java
@@ -1,0 +1,10 @@
+package hello.moviebook.book.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+@Controller
+@RequestMapping("/api/v1/book")
+@RequiredArgsConstructor
+public class BookController {
+}
