+++ movie-book/src/main/java/hello/moviebook/book/domain/Book.java
@@ -1,8 +1,7 @@
-package hello.moviebook.Book;
+package hello.moviebook.book.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import hello.moviebook.UserBook.UserBook;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
