+++ movie-book/src/main/java/hello/moviebook/book/service/BookService.java
@@ -1,0 +1,9 @@
+package hello.moviebook.book.service;
+
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+public class BookService {
+}
